requirements.txt
langchain==0.0.267
pandas==2.0.3
Pillow==9.5.0
Pillow==10.0.0
PyPDF2==3.0.1
python-dotenv==1.0.0
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.31.0
SpeechRecognition==3.10.0
streamlit==1.25.0
streamlit_authenticator==0.2.2
torch==2.0.1
transformers==4.31.0
PyAudio==0.2.13
openai==0.27.8